Fix to problem where data on NFS mounts is deleted when mounting nested volumes.

+++ netshare/drivers/nfs.go
@@ -119,9 +119,13 @@ func (n nfsDriver) Unmount(r volume.UnmountRequest) volume.Response {
 
 	n.mountm.DeleteIfNotManaged(resolvedName)
 
-	if err := os.RemoveAll(hostdir); err != nil {
-		return volume.Response{Err: err.Error()}
-	}
+        // Check if directory is empty. This command will return "err" if empty
+        if err := run(fmt.Sprintf("ls -1 %s | grep .", hostdir)); err == nil {
+                log.Warnf("Directory %s not empty after unmount. Skipping RemoveAll call.", hostdir)
+        } else {
+                if err := os.RemoveAll(hostdir); err != nil {
+                      return volume.Response{Err: err.Error()}
+        }
 
 	return volume.Response{}
 }
